fix(ai): remove use of `expect()` from production code
