+++ packages/ai/src/error/no-object-generated-error.ts
@@ -68,22 +68,3 @@ export class NoObjectGeneratedError extends AISDKError {
     return AISDKError.hasMarker(error, marker);
   }
 }
-
-export function verifyNoObjectGeneratedError(
-  error: unknown,
-  expected: {
-    message: string;
-    response: LanguageModelResponseMetadata & {
-      body?: string;
-    };
-    usage: LanguageModelUsage;
-    finishReason: FinishReason;
-  },
-) {
-  expect(NoObjectGeneratedError.isInstance(error)).toBeTruthy();
-  const noObjectGeneratedError = error as NoObjectGeneratedError;
-  expect(noObjectGeneratedError.message).toEqual(expected.message);
-  expect(noObjectGeneratedError.response).toEqual(expected.response);
-  expect(noObjectGeneratedError.usage).toEqual(expected.usage);
-  expect(noObjectGeneratedError.finishReason).toEqual(expected.finishReason);
-}
