+++ packages/ai/src/generate-text/output.test.ts
@@ -1,6 +1,6 @@
 import { fail } from 'assert';
 import { z } from 'zod/v4';
-import { verifyNoObjectGeneratedError } from '../error/no-object-generated-error';
+import { verifyNoObjectGeneratedError } from '../error/verify-no-object-generated-error';
 import { object } from './output';
 import { FinishReason } from '../types';
